requirements.txt
# Image processing
Pillow>=10.0.0